requirements.txt
aiogram==2.11.2
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==3.1
kiwisolver==1.3.1
matplotlib==3.4.0
multidict==5.1.0
numpy==1.20.1
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.2
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
